chore(SFINT-3596): Big refactor
